+++ tests/unit/base/test_commands_convert_onnx_onnx2trt.py
@@ -43,8 +43,9 @@ def test_run_execute_conversion_when_model_not_support_batching(mocker):
         output_model_path = workspace / "trt-fp16" / "model.plan"
         output_model_path.parent.mkdir(parents=True)
 
-        with mocker.patch.object(ConvertONNX2TRT, "_execute_conversion"), mocker.patch(
-            "model_navigator.utils.devices.get_available_gpus", return_value=[0]
+        with (
+            mocker.patch.object(ConvertONNX2TRT, "_execute_conversion"),
+            mocker.patch("model_navigator.utils.devices.get_available_gpus", return_value=[0]),
         ):
             result = ConvertONNX2TRT().run(
                 workspace=Workspace(workspace),
@@ -89,8 +90,9 @@ def test_run_execute_conversion_with_max_bs_search_when_trt_profile_not_provided
         output_model_path = workspace / "trt-fp16" / "model.plan"
         output_model_path.parent.mkdir(parents=True)
 
-        with mocker.patch.object(ConvertONNX2TRT, "_execute_conversion", return_value=1), mocker.patch(
-            "model_navigator.utils.devices.get_available_gpus", return_value=[0]
+        with (
+            mocker.patch.object(ConvertONNX2TRT, "_execute_conversion", return_value=1),
+            mocker.patch("model_navigator.utils.devices.get_available_gpus", return_value=[0]),
         ):
             result = ConvertONNX2TRT().run(
                 workspace=Workspace(workspace),
@@ -126,8 +128,9 @@ def test_run_execute_conversion_when_dataloader_and_device_max_batch_size_is_inv
         output_model_path = workspace / "trt-fp16" / "model.plan"
         output_model_path.parent.mkdir(parents=True)
 
-        with mocker.patch.object(ConvertONNX2TRT, "_execute_conversion", return_value=1), mocker.patch(
-            "model_navigator.utils.devices.get_available_gpus", return_value=[0]
+        with (
+            mocker.patch.object(ConvertONNX2TRT, "_execute_conversion", return_value=1),
+            mocker.patch("model_navigator.utils.devices.get_available_gpus", return_value=[0]),
         ):
             result = ConvertONNX2TRT().run(
                 workspace=Workspace(workspace),
@@ -165,8 +168,9 @@ def test_run_execute_single_conversion_when_only_dataloader_max_batch_size_provi
         output_model_path = workspace / "trt-fp16" / "model.plan"
         output_model_path.parent.mkdir(parents=True)
 
-        with mocker.patch.object(ConvertONNX2TRT, "_execute_single_conversion", return_value=3), mocker.patch(
-            "model_navigator.utils.devices.get_available_gpus", return_value=[0]
+        with (
+            mocker.patch.object(ConvertONNX2TRT, "_execute_single_conversion", return_value=3),
+            mocker.patch("model_navigator.utils.devices.get_available_gpus", return_value=[0]),
         ):
             result = ConvertONNX2TRT().run(
                 workspace=Workspace(workspace),
@@ -203,8 +207,9 @@ def test_run_execute_single_conversion_when_only_device_max_batch_size_provided(
         output_model_path = workspace / "trt-fp16" / "model.plan"
         output_model_path.parent.mkdir(parents=True)
 
-        with mocker.patch.object(ConvertONNX2TRT, "_execute_single_conversion", return_value=3), mocker.patch(
-            "model_navigator.utils.devices.get_available_gpus", return_value=[0]
+        with (
+            mocker.patch.object(ConvertONNX2TRT, "_execute_single_conversion", return_value=3),
+            mocker.patch("model_navigator.utils.devices.get_available_gpus", return_value=[0]),
         ):
             result = ConvertONNX2TRT().run(
                 workspace=Workspace(workspace),
@@ -241,9 +246,10 @@ def test_run_execute_conversion_with_max_batch_size_search_when_both_max_batch_s
         output_model_path = workspace / "trt-fp16" / "model.plan"
         output_model_path.parent.mkdir(parents=True)
 
-        with mocker.patch.object(
-            ConvertONNX2TRT, "_execute_conversion_with_max_batch_size_search", return_value=3
-        ), mocker.patch("model_navigator.utils.devices.get_available_gpus", return_value=[0]):
+        with (
+            mocker.patch.object(ConvertONNX2TRT, "_execute_conversion_with_max_batch_size_search", return_value=3),
+            mocker.patch("model_navigator.utils.devices.get_available_gpus", return_value=[0]),
+        ):
             result = ConvertONNX2TRT().run(
                 workspace=Workspace(workspace),
                 parent_path=input_model_path,